third_party/WebKit/LayoutTests/fast/mediacapturefromelement/HTMLMediaElementCapture-creation.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script>

// Run captureStream() on <video>/<audio>s and inspect the generated Stream.

test(function() {
  var video = document.createElement('video');
  assert_throws("NotSupportedError", function () { video.captureStream() },
                "captureStream() cannot be created out of a source-less <video>" );
}, 'check that captureStream() raises an exception on a <video> with no source.');

test(function() {
  var audio = document.createElement('audio');
  assert_throws("NotSupportedError", function () { audio.captureStream() },
                "captureStream() cannot be created out of a source-less <audio>" );
}, 'check that captureStream() raises an exception on an <audio> with no source.');

var makeAsyncTest = function(filename) {
  async_test(function() {
    var video = document.createElement('video');
    video.src = "../../http/tests/media/resources/media-source/webm/" + filename;
    video.onerror = this.unreached_func("<video> error");

    video.onloadstart = this.step_func_done(function() {
      var stream = video.captureStream();
      assert_not_equals(stream, null, "error generating stream");
    });
  }), "<video>.captureStream()";
};

generate_tests(makeAsyncTest,
               [[ "video-only", "test-v-128k-320x240-24fps-8kfr.webm"],
                [ "audio-only", "test-a-128k-44100Hz-1ch.webm"],
                [ "video+audio", "test.webm"]]);

</script>
